Providing initial documentation (wip) and file structure
